[Exercício] Invoices nível banco

+++ relations/database/migrations/2024_03_21_024535_create_invoices_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('description')->nullable();
             $table->float('value')->default(0);
-            $table->unsignedSmallInteger('address_id');
+            $table->unsignedBigInteger('address_id');
             $table->unsignedBigInteger('user_id');
             $table->timestamps();
         });
